+++ app_plus/src/main/java/org/gkisalatiga/plus/fragment/FragmentSeasonalMain.kt
@@ -51,39 +51,29 @@ import org.gkisalatiga.plus.screen.ScreenSeasonalCompanion
 class FragmentSeasonalMain (private val current : ActivityData) : ComponentActivity() {
 
     // The root seasonal node.
-    // private val seasonalNode = ModulesCompanion.jsonRoot!!.getJSONObject("seasonal")
     private val seasonalNode = ModulesCompanion.api!!.seasonal
 
     private val itemTargetTitle = listOf(
         seasonalNode.staticMenu.agenda.title,
         seasonalNode.staticMenu.books.title,
         seasonalNode.staticMenu.gallery.title,
+        seasonalNode.staticMenu.playlist.title,
         seasonalNode.staticMenu.twibbon.title,
     )
-    /*private val itemTargetTitle = listOf(
-        seasonalNode.getJSONObject("static-menu").getJSONObject("agenda").getString("title"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("books").getString("title"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("gallery").getString("title"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("twibbon").getString("title"),
-    )*/
 
     private val itemTargetBannerUrl = listOf(
         seasonalNode.staticMenu.agenda.banner,
         seasonalNode.staticMenu.books.banner,
         seasonalNode.staticMenu.gallery.banner,
+        seasonalNode.staticMenu.playlist.banner,
         seasonalNode.staticMenu.twibbon.banner,
     )
-    /*private val itemTargetBannerUrl = listOf(
-        seasonalNode.getJSONObject("static-menu").getJSONObject("agenda").getString("banner"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("books").getString("banner"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("gallery").getString("banner"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("twibbon").getString("banner"),
-    )*/
 
     private val itemTargetFragments = listOf(
         NavigationRoutes.FRAG_SEASONAL_AGENDA,
         NavigationRoutes.FRAG_SEASONAL_BOOK,
         NavigationRoutes.FRAG_SEASONAL_GALLERY,
+        NavigationRoutes.FRAG_SEASONAL_PLAYLIST,
         NavigationRoutes.FRAG_SEASONAL_TWIBBON,
     )
 
@@ -91,14 +81,9 @@ class FragmentSeasonalMain (private val current : ActivityData) : ComponentActiv
         seasonalNode.staticMenu.agenda.isShown,
         seasonalNode.staticMenu.books.isShown,
         seasonalNode.staticMenu.gallery.isShown,
+        seasonalNode.staticMenu.playlist.isShown,
         seasonalNode.staticMenu.twibbon.isShown,
     )
-    /*private val itemTargetIsShown = listOf(
-        seasonalNode.getJSONObject("static-menu").getJSONObject("agenda").getInt("is_shown"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("books").getInt("is_shown"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("gallery").getInt("is_shown"),
-        seasonalNode.getJSONObject("static-menu").getJSONObject("twibbon").getInt("is_shown"),
-    )*/
 
     @Composable
     fun getComposable() {